backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiohttp==3.9.1
aiofiles==23.2.1
python-multipart==0.0.6
Pillow==10.1.0
python-dateutil==2.8.2
